Renamed Clauses folder into Traits
